hospitals_project/core_project/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django
django-ninja==1.1.0
pydantic==2.7.1
pydantic_core==2.18.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
